feat: add atMost one, atLeast one and atLEast k
