+++ src/test/scala/satify/dsl/SatEncodingsTest.scala
@@ -1,12 +1,13 @@
 package satify.dsl
 
 import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.should.Matchers.shouldBe
+import org.scalatest.matchers.should.Matchers.{should, shouldBe}
 import satify.model.Expression
 
 class SatEncodingsTest extends AnyFlatSpec:
 
   import satify.dsl.DSL.{*, given}
+  import satify.dsl.Encodings.*
 
   """ atLeastOne("A", "B", "C")  """ should """ be equal to "A" or "B" or "C" """ in {
     atLeastOne("A", "B", "C") shouldBe ("A" or "B" or "C")
@@ -21,4 +22,24 @@ class SatEncodingsTest extends AnyFlatSpec:
       ((not("B") or "ENC1") and
         (not("ENC0") or "ENC1") and
         (not("B") or not("ENC0"))))
+  }
+
+  """ ("A", "B") atMost 1 """ should """ be equal to atMostOne("A", "B") """ in {
+    ("A", "B") atMost 1 shouldBe ((not("A") or "ENC0") and (not("B") or "ENC0"))
+  }
+
+  """ ("A", "B") atMost one """ should """ be equal to ("A", "B") atMost 1 """ in {
+    ("A", "B") atMost one shouldBe (("A", "B") atMost 1)
+  }
+
+  """ ("A", "B", "C") atLeast 1 """ should """ be equal to atLeastOne("A", "B") """ in {
+    ("A", "B", "C") atLeast 1 shouldBe ("A" or "B" or "C")
+  }
+
+  """ ("A", "B", "C") atLeast one """ should """ be equal to ("A", "B", "C") atLeast 1 """ in {
+    ("A", "B", "C") atLeast one shouldBe (("A", "B", "C") atLeast 1)
+  }
+
+  """ atLeastK(2, "A", "B", "C")  """ should """ be equal to "A" or "B" or "C" """ in {
+    atLeastK(2, "A", "B", "C") shouldBe (("A" and "B") or ("A" and "C") or ("B" and "C"))
   }
